requirements.txt
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tabulate==0.9.0
tzdata==2024.2
